requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
pytest
pytest-asyncio
gunicorn
tavily-python
mcp
openai